pyproject.toml
[tool.poetry]
name = "SaryahAI"
version = "0.1.0"
description = ""
authors = ["Khaild Alrasheed <Khaild@cycls.io>", "Umair Ahmed umair.ahmed@cycls.io"]

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
gunicorn = "^21.2.0"
urllib3 = "^1.26.7"
redis = "^5.0.1"
asyncpg = "0.28.0"
greenlet = "*"
psycopg2-binary = "^2.9.7"
python-socketio = "^5.10.0"

[tool.poetry.group.dev.dependencies]
black = { version = "^23.1.0", allow-prereleases = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
